+++ spec/lib/packaging/paths_spec.rb
@@ -5,11 +5,11 @@ describe 'Pkg::Paths' do
     arch_transformations = {
       ['pkg/el-8-x86_64/puppet-agent-6.9.0-1.el8.x86_64.rpm', 'el', '8'] => 'x86_64',
       ['pkg/el/8/puppet6/aarch64/puppet-agent-6.5.0.3094.g16b6fa6f-1.el8.aarch64.rpm', 'el', '8'] => 'aarch64',
-      ['artifacts/fedora/32/puppet6/x86_64/puppet-agent-6.9.0-1.fc32.x86_64.rpm', 'fedora', '32'] => 'x86_64',
-      ['pkg/ubuntu-16.04-amd64/puppet-agent_4.99.0-1xenial_amd64.deb', 'ubuntu', '16.04'] => 'amd64',
+      ['artifacts/fedora/36/puppet6/x86_64/puppet-agent-6.9.0-1.fc32.x86_64.rpm', 'fedora', '36'] => 'x86_64',
+      ['pkg/ubuntu-18.04-amd64/puppet-agent_4.99.0-1bionic_amd64.deb', 'ubuntu', '18.04'] => 'amd64',
       ['artifacts/deb/focal/puppet6/puppet-agent_6.5.0.3094.g16b6fa6f-1focal_arm64.deb', 'ubuntu', '20.04'] => 'aarch64',
 
-      ['artifacts/ubuntu-16.04-i386/puppetserver_5.0.1-0.1SNAPSHOT.2017.07.27T2346puppetlabs1.debian.tar.gz', 'ubuntu', '16.04'] => 'source',
+      ['artifacts/ubuntu-18.04-i386/puppetserver_5.0.1-0.1SNAPSHOT.2017.07.27T2346puppetlabs1.debian.tar.gz', 'ubuntu', '18.04'] => 'source',
       ['artifacts/el/6/PC1/SRPMS/puppetserver-5.0.1.master-0.1SNAPSHOT.2017.08.18T0951.el6.src.rpm', 'el', '6'] => 'SRPMS'
     }
     arch_transformations.each do |path_array, arch|
@@ -23,7 +23,7 @@ describe 'Pkg::Paths' do
   describe '#tag_from_artifact_path' do
     path_tranformations = {
       'pkg/el-7-x86_64/puppet-agent-5.5.22-1.el8.x86_64.rpm' => 'el-7-x86_64',
-      'pkg/ubuntu-20.04-amd64/puppet-agent_5.5.22-1xenial_amd64.deb' => 'ubuntu-20.04-amd64',
+      'pkg/ubuntu-20.04-amd64/puppet-agent_5.5.22-1bionic_amd64.deb' => 'ubuntu-20.04-amd64',
       'pkg/windows/puppet-agent-5.5.22-x86.msi' => 'windows-2012-x86',
       'artifacts/el/6/products/x86_64/pe-r10k-2.5.4.3-1.el6.x86_64.rpm' => 'el-6-x86_64',
       'pkg/pe/rpm/el-6-i386/pe-puppetserver-2017.3.0.3-1.el6.noarch.rpm' => 'el-6-i386',
@@ -35,7 +35,7 @@ describe 'Pkg::Paths' do
       'pkg/apple/11/puppet6/x86_64/puppet-agent-6.19.0-1.osx11.dmg' => 'osx-11-x86_64',
       'pkg/windows/puppet-agent-1.9.0-x86.msi' => 'windows-2012-x86',
       'pkg/pe/rpm/el-6-i386/pe-puppetserver-2017.3.0.3-1.el6.src.rpm' => 'el-6-SRPMS',
-      'pkg/pe/deb/xenial/pe-puppetserver_2017.3.0.3-1puppet1.orig.tar.gz' => 'ubuntu-16.04-source',
+      'pkg/pe/deb/bionic/pe-puppetserver_2017.3.0.3-1puppet1.orig.tar.gz' => 'ubuntu-18.04-source',
       'pkg/puppet-agent-5.1.0.79.g782e03c.gem' => nil,
       'pkg/puppet-agent-5.1.0.7.g782e03c.tar.gz' => nil,
     }
@@ -276,11 +276,11 @@ describe 'Pkg::Paths' do
         .to eq(fake_apt_repo_path)
     end
     it 'returns nonfinal_yum_repo_path for nonfinal rpms' do
-      expect(Pkg::Paths.remote_repo_base('fedora-34-x86_64', nonfinal: true))
+      expect(Pkg::Paths.remote_repo_base('fedora-36-x86_64', nonfinal: true))
         .to eq(fake_yum_nightly_repo_path)
     end
     it 'returns nonfinal_apt_repo_path for nonfinal debs' do
-      expect(Pkg::Paths.remote_repo_base('debian-9-amd64', nonfinal: true))
+      expect(Pkg::Paths.remote_repo_base('debian-10-amd64', nonfinal: true))
         .to eq(fake_apt_nightly_repo_path)
     end
     it 'fails if neither tag nor package_format is provided' do
@@ -309,8 +309,8 @@ describe 'Pkg::Paths' do
         allow(Pkg::Paths).to receive(:remote_repo_base).and_return('/opt/repository/apt')
         expect(Pkg::Paths.apt_package_base_path('ubuntu-18.04-amd64', 'puppet6', 'puppet-agent'))
           .to eq('/opt/repository/apt/pool/bionic/puppet6/p/puppet-agent')
-        expect(Pkg::Paths.apt_package_base_path('debian-9-amd64', 'puppet6', 'bolt-server'))
-          .to eq('/opt/repository/apt/pool/stretch/puppet6/b/bolt-server')
+        expect(Pkg::Paths.apt_package_base_path('debian-10-amd64', 'puppet6', 'bolt-server'))
+          .to eq('/opt/repository/apt/pool/buster/puppet6/b/bolt-server')
 
 
       end
@@ -364,16 +364,16 @@ describe 'Pkg::Paths' do
           .to eq("#{yum_repo_path}/#{repo_name}-release-sles-12.noarch.rpm")
       end
       it 'returns the appropriate link path for deb release packages' do
-        expect(Pkg::Paths.release_package_link_path('ubuntu-16.04-amd64'))
-          .to eq("#{apt_repo_path}/#{repo_name}-release-xenial.deb")
+        expect(Pkg::Paths.release_package_link_path('ubuntu-18.04-amd64'))
+          .to eq("#{apt_repo_path}/#{repo_name}-release-bionic.deb")
       end
       it 'returns the appropriate link path for nonfinal rpm release packages' do
         expect(Pkg::Paths.release_package_link_path('el-7-x86_64', true))
           .to eq("#{nonfinal_yum_repo_path}/#{nonfinal_repo_name}-release-el-7.noarch.rpm")
       end
       it 'returns the appropriate link path for nonfinal deb release packages' do
-        expect(Pkg::Paths.release_package_link_path('debian-9-i386', true))
-          .to eq("#{nonfinal_apt_repo_path}/#{nonfinal_repo_name}-release-stretch.deb")
+        expect(Pkg::Paths.release_package_link_path('debian-10-amd64', true))
+          .to eq("#{nonfinal_apt_repo_path}/#{nonfinal_repo_name}-release-buster.deb")
       end
       it 'returns nil for package formats that do not have release packages' do
         expect(Pkg::Paths.release_package_link_path('osx-10.15-x86_64')).to eq(nil)